refactor(config-ui): remove the pages/pipeline/detail
