+++ config-ui/src/pages/pipeline/api.ts
@@ -18,6 +18,25 @@
 
 import { request } from '@/utils';
 
+export const getPipeline = (id: ID) => request(`/pipelines/${id}`);
+
+export const getPipelineTasks = (id: ID) => request(`/pipelines/${id}/tasks`);
+
+export const deletePipeline = (id: ID) =>
+  request(`/pipelines/${id}`, {
+    method: 'delete',
+  });
+
 export const getPipelineHistorical = (id: ID) => request(`/blueprints/${id}/pipelines`);
+
+export const pipelineRerun = (id: ID) =>
+  request(`/pipelines/${id}/rerun`, {
+    method: 'post',
+  });
+
+export const taskRerun = (id: ID) =>
+  request(`/tasks/${id}/rerun`, {
+    method: 'post',
+  });
 
 export const getPipelineLog = (id: ID) => request(`/pipelines/${id}/logging.tar.gz`);
