--- config-ui/src/pages/pipeline/detail/use-detail.ts
@@ -1,73 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-import { useState, useMemo } from 'react';
-
-import { operator } from '@/utils';
-
-import * as API from './api';
-
-export interface Props {
-  id: ID;
-}
-
-export const useDetail = ({ id }: Props) => {
-  const [version, setVersion] = useState(0);
-  const [operating, setOperating] = useState(false);
-
-  const handlePipelineCancel = async () => {
-    const [success] = await operator(() => API.deletePipeline(id), {
-      setOperating,
-    });
-
-    if (success) {
-      setVersion(version + 1);
-    }
-  };
-
-  const handlePipelineRerun = async () => {
-    const [success] = await operator(() => API.pipelineRerun(id), {
-      setOperating,
-    });
-
-    if (success) {
-      setVersion(version + 1);
-    }
-  };
-
-  const handleTaskRertun = async (id: ID) => {
-    const [success] = await operator(() => API.taskRerun(id), {
-      setOperating,
-    });
-
-    if (success) {
-      setVersion(version + 1);
-    }
-  };
-
-  return useMemo(
-    () => ({
-      version,
-      operating,
-      onCancel: handlePipelineCancel,
-      onRerun: handlePipelineRerun,
-      onRerunTask: handleTaskRertun,
-    }),
-    [version, operating],
-  );
-};
