+++ config-ui/src/pages/pipeline/index.ts
@@ -17,6 +17,4 @@
  */
 
 export * from './types';
-export * from './misc';
 export * from './components';
-export * from './detail';
